+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 - [TaxonKit v0.14.2](https://github.com/shenwei356/taxonkit/releases/tag/v0.14.2)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/taxonkit/v0.14.2/total.svg)](https://github.com/shenwei356/taxonkit/releases/tag/v0.14.2)
+    - `taxonkit filter`:
+        - fix checking merged/deleted/not-found taxids. [#80](https://github.com/shenwei356/taxonkit/issues/80)
     - `taxonkit lca`:
         - add a new flag `-b/--buffer-size` to set the size of the line buffer. [#75](https://github.com/shenwei356/taxonkit/issues/75)
         - fix typo: `--separater` -> `--separater`, the former is still available for backward compatibility.